+++ src/watts/plugin_pyarc.py
@@ -73,9 +73,8 @@ class PluginPyARC(PluginGeneric):
     ):
         executable = _find_executable(executable, 'PyARC_DIR')
         super().__init__(executable, None, template_file, extra_inputs,
-                         extra_template_inputs, show_stdout, show_stderr)
+                         extra_template_inputs, "PyARC", show_stdout, show_stderr)
         self.input_name = "pyarc_input.son"
-        self.plugin_name = "PyARC"
 
     @PluginGeneric.executable.setter
     def executable(self, exe: PathLike):